feat: マウスホイールズーム＆左クリックドラッグパンを追加
ワールドビューポート上でのマウスホイールズームと、左クリック＋ドラッグによる
カメラパン操作を実装。ドラッグ閾値判定により誤クリックを防止する。

+++ game-plugin/src/resources/interaction.rs
@@ -4,7 +4,7 @@ use bevy::prelude::*;
 
 /// 現在マウスがホバーしているセルの座標を保持するリソース
 #[derive(Resource, Default, PartialEq)]
-pub struct HoveredCell(pub Option<(u16, u16)>);
+pub struct HoveredCell(pub Option<(i32, i32)>);
 
 /// 連続再生防止用のクールダウンタイマー（50ms）
 /// TimerMode::Once なので reset() 後に再度50ms経過するまで is_finished() は false を返す
